Rename: components, pages 폴더 분리 및 파일 이동
- Home 컴포넌트, Simulator 컴포넌트 파일 이동(components -> pages)
- SimulCanvas, SimulController, Chuck 파일 이동(components/Simulation -> components)
- Header 파일 이동(components/Header -> components)
