+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from "react-router-dom";
-import Home from "./components/Home/Home";
-import Simulator from "./components/Simulation/Simulator";
+import Home from "./pages/Home";
+import Simulator from "./pages/Simulator";
 
 const App = () => {
   return (
