+++ src/components/SimulCanvas.jsx
